refactor(): moved old main client model code to model_old - created updated/cleaned up based model
